requirements.txt
python-docx
scikit-learn
transformers
torch>=1.10.0
spacy
nltk
tabulate